+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implemented Webhook improvements [issue 140](https://github.com/Comcast/caduceus/issues/140)
+
 ### Changed
-- retry on non 2xx status codes
+- Retry on non 2xx status codes
 
 ## [0.1.3] - 2019-02-27
 ### Changed